backend/requirements.txt
Flask
firebase-admin
python-dotenv
gtts
SpeechRecognition
assemblyai
openai